src/main/java/com/drkiettran/mapreduce/IntSumReducer.java
package com.drkiettran.mapreduce;

import java.io.IOException;

import org.apache.hadoop.io.IntWritable;
import org.apache.hadoop.io.Text;
import org.apache.hadoop.mapreduce.Reducer;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

public class IntSumReducer extends Reducer<Text, IntWritable, Text, IntWritable> {
	private static final Logger LOGGER = LoggerFactory.getLogger(IntSumReducer.class);
	private IntWritable result = new IntWritable();

	@Override
	public void reduce(Text key, Iterable<IntWritable> values, Context context)
			throws IOException, InterruptedException {
		int sum = 0;
		for (IntWritable val : values) {
			sum += val.get();
		}
		result.set(sum);
		context.write(key, result);
		LOGGER.info("{} occurs {} times", key, sum);
		System.out.println(String.format("%s occurs %d times", key, sum));
	}
}
